agent-builder-agent/pyproject.toml
[project]
name = "agent-builder-agent"
version = "0.1.0"
description = "Autonomous AI agent that builds, documents, and ships one new AI agent every day"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.9.0",
    "langgraph>=0.2.0",
    "langchain-openai>=0.2.0",
    "langchain-core>=0.3.0",
    "python-dotenv>=1.2.1",
    "schedule>=1.2.0",
]

